refactor: ♻️ Auth 패키지에서 User 도메인 분리 (#30)
- Member, Gender, MemberStatus를 auth/domain → user/domain으로 이동
- MemberRepository를 auth/repository → user/repository로 이동
- 전체 참조 파일(25개) import 경로 일괄 수정
