--- src/main/java/com/doori/doori_backend/auth/domain/Gender.java
@@ -1,5 +1,0 @@
-package com.doori.doori_backend.auth.domain;
-
-public enum Gender {
-    M, F
-}
