+++ src/main/java/com/doori/doori_backend/auth/dto/request/SignupRequest.java
@@ -1,6 +1,6 @@
 package com.doori.doori_backend.auth.dto.request;
 
-import com.doori.doori_backend.auth.domain.Gender;
+import com.doori.doori_backend.user.domain.Gender;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
