+++ src/main/java/com/doori/doori_backend/block/domain/Block.java
@@ -1,6 +1,6 @@
 package com.doori.doori_backend.block.domain;
 
-import com.doori.doori_backend.auth.domain.Member;
+import com.doori.doori_backend.user.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
