+++ src/main/java/com/doori/doori_backend/block/dto/response/BlockedUserResponse.java
@@ -1,6 +1,6 @@
 package com.doori.doori_backend.block.dto.response;
 
-import com.doori.doori_backend.auth.domain.Member;
+import com.doori.doori_backend.user.domain.Member;
 
 public record BlockedUserResponse(
     Long userId,
