+++ src/main/java/com/doori/doori_backend/block/service/BlockService.java
@@ -1,8 +1,8 @@
 package com.doori.doori_backend.block.service;
 
-import com.doori.doori_backend.auth.domain.Member;
-import com.doori.doori_backend.auth.domain.MemberStatus;
-import com.doori.doori_backend.auth.repository.MemberRepository;
+import com.doori.doori_backend.user.domain.Member;
+import com.doori.doori_backend.user.domain.MemberStatus;
+import com.doori.doori_backend.user.repository.MemberRepository;
 import com.doori.doori_backend.block.domain.Block;
 import com.doori.doori_backend.block.dto.response.BlockedUserResponse;
 import com.doori.doori_backend.block.repository.BlockRepository;
