+++ src/main/java/com/doori/doori_backend/favorite/service/FavoriteService.java
@@ -1,8 +1,8 @@
 package com.doori.doori_backend.favorite.service;
 
-import com.doori.doori_backend.auth.domain.Member;
-import com.doori.doori_backend.auth.domain.MemberStatus;
-import com.doori.doori_backend.auth.repository.MemberRepository;
+import com.doori.doori_backend.user.domain.Member;
+import com.doori.doori_backend.user.domain.MemberStatus;
+import com.doori.doori_backend.user.repository.MemberRepository;
 import com.doori.doori_backend.favorite.domain.PostFavorite;
 import com.doori.doori_backend.favorite.domain.UserFavorite;
 import com.doori.doori_backend.favorite.dto.response.FavoritePostResponse;
