+++ src/main/java/com/doori/doori_backend/lifestyle/repository/LifestyleProfileRepository.java
@@ -1,7 +1,7 @@
 package com.doori.doori_backend.lifestyle.repository;
 
-import com.doori.doori_backend.auth.domain.Member;
-import com.doori.doori_backend.auth.domain.MemberStatus;
+import com.doori.doori_backend.user.domain.Member;
+import com.doori.doori_backend.user.domain.MemberStatus;
 import com.doori.doori_backend.lifestyle.domain.HousingType;
 import com.doori.doori_backend.lifestyle.domain.LifestyleProfile;
 import java.util.List;
