+++ src/main/java/com/doori/doori_backend/notification/domain/Notification.java
@@ -1,6 +1,6 @@
 package com.doori.doori_backend.notification.domain;
 
-import com.doori.doori_backend.auth.domain.Member;
+import com.doori.doori_backend.user.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
