+++ src/main/java/com/doori/doori_backend/user/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.doori.doori_backend.user.domain;
+
+public enum Gender {
+    M, F
+}
