+++ src/main/java/com/doori/doori_backend/user/dto/response/MyProfileResponse.java
@@ -1,6 +1,6 @@
 package com.doori.doori_backend.user.dto.response;
 
-import com.doori.doori_backend.auth.domain.Member;
+import com.doori.doori_backend.user.domain.Member;
 import java.time.LocalDateTime;
 
 public record MyProfileResponse(
