+++ src/main/java/com/doori/doori_backend/user/service/UserProfileService.java
@@ -1,8 +1,8 @@
 package com.doori.doori_backend.user.service;
 
-import com.doori.doori_backend.auth.domain.Member;
-import com.doori.doori_backend.auth.domain.MemberStatus;
-import com.doori.doori_backend.auth.repository.MemberRepository;
+import com.doori.doori_backend.user.domain.Member;
+import com.doori.doori_backend.user.domain.MemberStatus;
+import com.doori.doori_backend.user.repository.MemberRepository;
 import com.doori.doori_backend.auth.repository.RefreshTokenRepository;
 import com.doori.doori_backend.global.error.ErrorCode;
 import com.doori.doori_backend.global.exception.CustomException;
